requirements.txt
telebot~=0.0.5
python-dotenv~=1.2.2
gspread~=6.0
google-auth~=2.40